+++ CHANGELOG.md
@@ -6,7 +6,13 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- disco: add `Hash` and `AppendHash` methods to the `Info` type for calculating
+  [XEP-0115: Entity Capabilities] hashes
 - xmpp: ability to create informational-only stream features
+
+
+[XEP-0115: Entity Capabilities]: https://xmpp.org/extensions/xep-0115.html
+
 
 ## v0.21.1 — 2022-02-11